feat(analytics): support decimal stats and remove K-Means parameters from UI
- Convert player_stats columns (goals, assists, shots) from INTEGER to REAL type to support decimal values
- Update database constraints to validate decimal stats appropriately
- Remove K and maxIter parameters from analytics filter bar UI
- Hardcode K-Means parameters (k=3, maxIter=100) in analytics page logic
- Update player stats dialog validation to accept decimal inputs
- Simplify analytics filter layout from 4 to 3 columns
- Update analytics description to reflect fixed 3-cluster configuration
- Improve form validation and API endpoints to handle decimal stat values

+++ src/app/api/player-stats/[id]/route.ts
@@ -21,9 +21,9 @@ const paramsSchema = z.object({
 
 const updateSchema = z
   .object({
-    goals: z.coerce.number().int().min(0).optional(),
-    assists: z.coerce.number().int().min(0).optional(),
-    shots: z.coerce.number().int().min(0).optional(),
+    goals: z.coerce.number().min(0).optional(),
+    assists: z.coerce.number().min(0).optional(),
+    shots: z.coerce.number().min(0).optional(),
   })
   .refine((value) => Object.keys(value).length > 0, {
     message: "Minimal 1 field harus diisi",
